+++ app/helpers/admin/citations_helper.rb
@@ -3,16 +3,24 @@ module Admin::CitationsHelper
   ICONS = {
     journalism: '🗞️',
     campaigning: '📣',
-    academic: '🎓',
+    research: '📚',
     other: '🌐'
   }.with_indifferent_access.freeze
 
+  def citation_title(citation)
+    citation.title.presence || citation.source_url
+  end
+
   def citation_icon(citation)
+    citation_icon_for_type(citation.type)
+  end
+
+  def citation_icon_for_type(type)
     html_attrs = {
-      title: citation.type.humanize,
-      class: "citation-icon citation-icon--#{citation.type}"
+      title: type.humanize,
+      class: "citation-icon citation-icon--#{type}"
     }
 
-    tag.span(ICONS.fetch(citation.type), **html_attrs)
+    tag.span(ICONS.fetch(type), **html_attrs)
   end
 end