+++ config/routes.rb
@@ -173,18 +173,29 @@ Rails.application.routes.draw do
   end
   ####
 
-  scope path: 'request/:url_title' do
-    #### Citations controller
-    resources :citations, only: [:new, :create],
-      defaults: { resource: 'InfoRequest' }
-    ####
+  #### Citations controller
+  resources :citations, only: [:index]
 
-    #### Classifications controller
+  scope path: 'request/:url_title' do
+    resources :citations,
+      only: [:new, :create],
+      defaults: { resource: 'InfoRequest' }
+  end
+
+  resources :info_request_batch, :only => :show do
+    resources :citations,
+      only: [:new, :create],
+      defaults: { resource: 'InfoRequestBatch' }
+  end
+  ####
+
+  #### Classifications controller
+  scope path: 'request/:url_title' do
     resources :classifications, only: [:create], param: :described_state do
       get :message, on: :member
     end
-    ####
   end
+  ####
 
   #### Followups controller
   match '/request/:request_url_title/followups/new' => 'followups#new',
@@ -235,13 +246,6 @@ Rails.application.routes.draw do
     resource :report, :only => [:new, :create]
     resource :widget, :only => [:new, :show]
     resources :widget_votes, :only => [:create]
-  end
-
-  resources :info_request_batch, :only => :show do
-    #### Citations controller
-    resources :citations, only: [:new, :create],
-      defaults: { resource: 'InfoRequestBatch' }
-    ####
   end
 
   #### OutgoingMessage controller
@@ -528,7 +532,7 @@ Rails.application.routes.draw do
 
   #### Admin::Citations controller
   namespace :admin do
-    resources :citations, only: [:index]
+    resources :citations, only: [:index, :edit, :update, :destroy]
   end
   ####