+++ db/migrate/20241007090524_add_title_and_description_to_citations.rb
@@ -1,0 +1,6 @@
+class AddTitleAndDescriptionToCitations < ActiveRecord::Migration[7.0]
+  def change
+    add_column :citations, :title, :string
+    add_column :citations, :description, :text
+  end
+end