+++ spec/views/info_request_batch/_citations.html.erb_spec.rb
@@ -34,7 +34,7 @@ RSpec.describe 'info_request_batch/citations' do
       before { render_view }
 
       it 'renders the section' do
-        expect(rendered).to match(/In the News/)
+        expect(rendered).to match(/FOI in Action/)
       end
 
       it 'renders the blank slate text' do
@@ -59,11 +59,7 @@ RSpec.describe 'info_request_batch/citations' do
       before { render_view }
 
       it 'renders the section' do
-        expect(rendered).to match(/In the News/)
-      end
-
-      it 'does not render the blank slate text' do
-        expect(rendered).not_to match(/Has this batch request been referenced/)
+        expect(rendered).to match(/FOI in Action/)
       end
 
       it 'renders the citations' do
@@ -81,19 +77,19 @@ RSpec.describe 'info_request_batch/citations' do
       before { render_view }
 
       it 'renders the section' do
-        expect(rendered).to match(/In the News/)
-      end
-
-      it 'does not render the blank slate text' do
-        expect(rendered).not_to match(/Has this request been referenced/)
+        expect(rendered).to match(/FOI in Action/)
       end
 
       it 'renders the citations' do
         expect(rendered).to match(/citations-list/)
       end
 
+      it 'renders the blank slate text' do
+        expect(rendered).to match(/Has this batch request been referenced/)
+      end
+
       it 'renders the link to add citations' do
-        expect(rendered).to match('New Citation')
+        expect(rendered).to match('Let us know')
       end
     end
   end