+++ doc/CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Highlighted Features
 
+* Improve citations admin to allow title and description updates (Graeme
+  Porteous)
+* Add public citations view (Graeme Porteous)
 * Add project dataset question "select" format (Graeme Porteous)
 * Fix script/mailin when multiple EXCEPTION_NOTIFICATIONS_TO addresses are
   specified (Graeme Porteous)
@@ -33,7 +36,7 @@
 * Strengthen 2FA warning. Users *must* remember to keep this code safe (Gareth
   Rees)
 * Broaden citation type classifications to cover wider thematic areas and add
-  "campaigning" type (Gareth Rees)
+  "campaigning" type (Gareth Rees, Graeme Porteous)
 * Treat unstemmed words as normal search terms. (Matthew Somerville)
 * Update `/alaveteli_pro` base path to `/pro` (Alexander Griffen, Graeme
   Porteous)